Move call_abi_checker to the only file using it
